source/Quart/requirementsQUART.txt
yt_dlp==2023.11.14
python-dotenv==1.0.0
requests>=2.28.0
quart>=0.18.0
quart-discord>=2.1.4
discord.py>=2.0.0
aiohttp>=3.8.0
hypercorn==0.14.3 
quart-wtforms>=0.4.0
werkzeug==2.2.0